+++ components/esp_lvgl_port/examples/touchscreen/main/main.c
@@ -182,7 +182,7 @@ static esp_err_t app_lvgl_init(void)
         .hres = EXAMPLE_LCD_H_RES,
         .vres = EXAMPLE_LCD_V_RES,
         .monochrome = false,
-        /* Rotation values must be same as used in esp_lcd for initial settings of the screen */
+        .color_format = LV_COLOR_FORMAT_RGB565,
         .rotation = {
             .swap_xy = false,
             .mirror_x = true,